Set database ID on LifecycleEvent payload in DoctrineProvider
The Provider that inserts an audit log into a database should
add an ID information to payload so other providers can use that
information.

Let's say we have the provider that fills ElasticSearch, we
need to have the ID of database entry so afterward we can easily
find that entry no matter if we are looking in a database or
ElasticSearch.

Let's say we have the provider that fills ElasticSearch, we need to have the ID of database entry so afterward we can easily
find that entry no matter if we are looking in a database or ElasticSearch.